// src/components/rim/RimMain.types.ts
export interface DataRim {
  info: Info;
  results: Result[];
}

export interface Info {
  count: number;
  pages: number;
  next: string | null;
  prev: string | null;
}

export interface Result {
  id: number;
  name: string;
  status: string;
  species: string;
  type: string;
  gender: string;
  origin: {
    name: string;
    url: string;
  };
  location: {
    name: string;
    url: string;
  };
  image: string;
  episode: string[];
  url: string;
  created: string;
}

export type Status =
  | "load"
  | "ready"
  | "error"
  | "notFound"
  | "";